Added a modular, functional, approach to the getList function. Added src/utils/generate.js for this purpose
