+++ src/api/components/ability/ability-controller.js
@@ -1,5 +1,5 @@
 const service = require('./ability-service');
-const { errorResponder, errorTypes } = require('../../../core/errors');
+const {generateGetListFunc} = require('../../../utils/generate');
 
 async function getBy(request, response, next) {
     try {
@@ -13,45 +13,7 @@ async function getBy(request, response, next) {
     }
 }
 
-async function getList(request, response, next) {
-    try {
-        let { offset, limit } = request.query;
-        for (const i of [offset, limit]) {
-            if (i === undefined) continue;
-            const num = Number(i);
-            if (Number.isNaN(num)) {
-                throw errorResponder(
-                    errorTypes.ARGUMENT_TYPE,
-                    'Query arguments must be numbers!'
-                );
-            }
-            if (num < 0) {
-                throw errorResponder(
-                    errorTypes.ARGUMENT_TYPE,
-                    `Query arguments cannot be negative!`
-                );
-            }
-            if (!Number.isInteger(num)) {
-                throw errorResponder(
-                    errorTypes.ARGUMENT_TYPE,
-                    'Query arguments must be integers!'
-                );
-            }
-        }
-        offset = Number(offset) || 0;
-        limit = Number(limit) || 20;
-
-        const doc = await service.getList(offset, limit);
-
-        return response.status(200).json({
-            offset,
-            limit,
-            data: doc,
-        });
-    } catch (error) {
-        return next(error);
-    }
-}
+const getList = generateGetListFunc(service.getList);
 
 module.exports = {
     getBy,
