+++ src/api/components/berry/berry-controller.js
@@ -1,5 +1,5 @@
 const service = require('./berry-service');
-const { errorResponder, errorTypes } = require('../../../core/errors');
+const {generateGetListFunc} = require('../../../utils/generate');
 
 async function get(request, response, next) {
     try {
@@ -13,29 +13,7 @@ async function get(request, response, next) {
     }
 }
 
-async function getList(request, response, next) {
-    try {
-        const offset = Number(request.query.offset) || 0;
-        const limit = Number(request.query.limit) || 20;
-
-        if (offset < 0 || limit < 0) {
-            throw errorResponder(
-                errorTypes.ARGUMENT_TYPE,
-                `Query arguments cannot be negative!`
-            );
-        }
-
-        const doc = await service.getList(offset, limit);
-
-        return response.status(200).json({
-            offset,
-            limit,
-            data: doc,
-        });
-    } catch (error) {
-        return next(error);
-    }
-}
+const getList = generateGetListFunc(service.getList);
 
 module.exports = {
     get,
