+++ src/api/components/language/language-controller.js
@@ -13,29 +13,33 @@ async function get(request, response, next) {
     }
 }
 
-async function get(request, response, next) {
-    try {
-        const { str } = request.params;
-
-        const postedDocument = await service.getItem(str);
-
-        return response.status(200).json(postedDocument);
-    } catch (error) {
-        return next(error);
-    }
-}
-
 async function getList(request, response, next) {
     try {
-        const offset = Number(request.query.offset) || 0;
-        const limit = Number(request.query.limit) || 20;
-
-        if (offset < 0 || limit < 0) {
-            throw errorResponder(
-                errorTypes.ARGUMENT_TYPE,
-                `Query arguments cannot be negative!`
-            );
+        let { offset, limit } = request.query;
+        for (const i of [offset, limit]) {
+            if (i === undefined) continue;
+            const num = Number(i);
+            if (Number.isNaN(num)) {
+                throw errorResponder(
+                    errorTypes.ARGUMENT_TYPE,
+                    'Query arguments must be numbers!'
+                );
+            }
+            if (num < 0) {
+                throw errorResponder(
+                    errorTypes.ARGUMENT_TYPE,
+                    `Query arguments cannot be negative!`
+                );
+            }
+            if (!Number.isInteger(num)) {
+                throw errorResponder(
+                    errorTypes.ARGUMENT_TYPE,
+                    'Query arguments must be integers!'
+                );
+            }
         }
+        offset = Number(offset) || 0;
+        limit = Number(limit) || 20;
 
         const doc = await service.getList(offset, limit);
